Refactor for merge
