+++ AC-Group-UI/src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { AuthService } from './Services/auth/auth.service';
+import { AuthService } from './Services/auth.service';
 import { Component } from '@angular/core';
 import { AngularFirestore } from '@angular/fire/firestore';
 import { Observable } from 'rxjs';
